Extract `nat46` module
